Added background color and border color to table visitTypes, fixed exception on Server "en-UK is not supported culture info"   and some other small fixes

+++ PatientManagement/PatientManagement.Web/Modules/PatientManagement/Patients/PatientsColumns.cs
@@ -17,8 +17,9 @@ namespace PatientManagement.PatientManagement.Columns
         public Int32 PatientId { get; set; }
         [EditLink]
         public String Name { get; set; }
-        public Int32 PersonalNumber { get; set; }
-        public Int32 PhoneNumber { get; set; }
+
+        public Int64 PersonalNumber { get; set; }
+        public Int64 PhoneNumber { get; set; }
         public DateTime FirstRegistrationDate { get; set; }
         public String Address { get; set; }
         public Int32 Height { get; set; }
