+++ PatientManagement/PatientManagement.Web/Modules/PatientManagement/Patients/PatientsForm.cs
@@ -18,8 +18,10 @@ namespace PatientManagement.PatientManagement.Forms
 
       
         public String Name { get; set; }
-        public Int32 PersonalNumber { get; set; }
-        public Int32 PhoneNumber { get; set; }
+
+        public Int64 PersonalNumber { get; set; }
+        public Int64 PhoneNumber { get; set; }
+
         [DefaultValue("now")]
         public DateTime FirstRegistrationDate { get; set; }
         public String Address { get; set; }
